+++ auditlog/admin.py
@@ -1,7 +1,8 @@
 from django.contrib import admin
 from django.contrib.auth import get_user_model
+from django.utils.translation import gettext_lazy as _
 
-from auditlog.filters import ResourceTypeFilter
+from auditlog.filters import CIDFilter, ResourceTypeFilter
 from auditlog.mixins import LogEntryAdminMixin
 from auditlog.models import LogEntry
 
@@ -9,7 +10,14 @@ from auditlog.models import LogEntry
 @admin.register(LogEntry)
 class LogEntryAdmin(admin.ModelAdmin, LogEntryAdminMixin):
     list_select_related = ["content_type", "actor"]
-    list_display = ["created", "resource_url", "action", "msg_short", "user_url"]
+    list_display = [
+        "created",
+        "resource_url",
+        "action",
+        "msg_short",
+        "user_url",
+        "cid_url",
+    ]
     search_fields = [
         "timestamp",
         "object_repr",
@@ -18,13 +26,17 @@ class LogEntryAdmin(admin.ModelAdmin, LogEntryAdminMixin):
         "actor__last_name",
         f"actor__{get_user_model().USERNAME_FIELD}",
     ]
-    list_filter = ["action", ResourceTypeFilter]
+    list_filter = ["action", ResourceTypeFilter, CIDFilter]
     readonly_fields = ["created", "resource_url", "action", "user_url", "msg"]
     fieldsets = [
-        (None, {"fields": ["created", "user_url", "resource_url"]}),
-        ("Changes", {"fields": ["action", "msg"]}),
+        (None, {"fields": ["created", "user_url", "resource_url", "cid"]}),
+        (_("Changes"), {"fields": ["action", "msg"]}),
     ]
 
     def has_add_permission(self, request):
         # As audit admin doesn't allow log creation from admin
         return False
+
+    def get_queryset(self, request):
+        self.request = request
+        return super().get_queryset(request=request)