+++ auditlog_tests/models.py
@@ -1,6 +1,7 @@
 import uuid
 
 from django.contrib.postgres.fields import ArrayField
+from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models
 
 from auditlog.models import AuditlogHistoryField
@@ -21,6 +22,9 @@ class SimpleModel(models.Model):
     datetime = models.DateTimeField(auto_now=True)
 
     history = AuditlogHistoryField()
+
+    def __str__(self):
+        return self.text
 
 
 class AltPrimaryKeyModel(models.Model):
@@ -257,7 +261,7 @@ class NoDeleteHistoryModel(models.Model):
 
 
 class JSONModel(models.Model):
-    json = models.JSONField(default=dict)
+    json = models.JSONField(default=dict, encoder=DjangoJSONEncoder)
 
     history = AuditlogHistoryField(delete_related=False)